requirements.txt
asn1crypto>=0.24.0
certifi>=2019.6.16
chardet>=3.0.4
cryptography>=2.1.4
enum34>=1.1.6
future>=0.17.1
idna>=2.8
ipaddress>=1.0.17
iso8601>=0.1.12
keyring>=10.6.0
keyrings.alt>=3.0
mysql-connector>=2.2.9
mysql-connector-python>=2.1.6
oauthlib>=3.1.0
playsound>=1.2.2
pycrypto>=2.6.1
pygame>=1.9.6
pygobject>=3.26.1
pyserial>=3.4
python-dateutil>=2.8.0
python-twitter>=3.5
pyttsx3>=2.7
pyxdg>=0.25
PyYAML>=5.1.2
requests>=2.22.0
requests-oauthlib>=1.2.0
SecretStorage>=2.3.1
six>=1.12.0
termcolor>=1.1.0
urllib3>=1.25.3
